agents/requirements.txt
# Web scraping dependencies
playwright==1.48.0
selenium==4.27.1
beautifulsoup4==4.12.3
requests==2.32.3
aiohttp==3.9.1

# AI processing
openai==1.58.1
anthropic==0.40.0

# Data processing
pandas==2.0.3
numpy==1.24.3

# Scheduling and utilities
schedule==1.2.2
python-dotenv==1.0.1
pydantic==2.10.4

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0# Mo
nitoring and error tracking
sentry-sdk==2.34.1